Rename jog_arm_parameters.h file to servo_parameters.h and fix relevant includes
