+++ moveit_ros/moveit_servo/include/moveit_servo/collision_check.h
@@ -45,7 +45,7 @@
 #include <sensor_msgs/JointState.h>
 #include <std_msgs/Float64.h>
 
-#include <moveit_servo/jog_arm_parameters.h>
+#include <moveit_servo/servo_parameters.h>
 #include <moveit_servo/low_pass_filter.h>
 #include <moveit_servo/joint_state_subscriber.h>
