+++ moveit_ros/moveit_servo/include/moveit_servo/jog_arm.h
@@ -41,7 +41,7 @@
 #include <memory>
 
 #include <moveit_servo/collision_check.h>
-#include <moveit_servo/jog_arm_parameters.h>
+#include <moveit_servo/servo_parameters.h>
 #include <moveit_servo/jog_calcs.h>
 #include <moveit_servo/joint_state_subscriber.h>
